Cambios minimos

+++ src/vista/botones/BotonAccion.java
@@ -27,8 +27,8 @@ public class BotonAccion extends StackPane{
 
 	private VBox boton;
 	private LibreriaSonidos sonidos;
-	public static int anchoBotonAccion = (int) (Screen.getPrimary().getVisualBounds().getWidth()/8); //cambio  200
-	public static int altoBotonAccion = (int) (Screen.getPrimary().getVisualBounds().getHeight()/8.5); //cambio  120
+	public static double anchoBotonAccion = Screen.getPrimary().getVisualBounds().getWidth() / 8 ; //cambio  200
+	public static double altoBotonAccion = Screen.getPrimary().getVisualBounds().getHeight() / 8.5; //cambio  120
 	public static double tamFuente = Screen.getPrimary().getVisualBounds().getHeight() / 42 ;
